+++ lmoments3/tests/test_lmoments.py
@@ -1,5 +1,6 @@
 import unittest
 
+import numpy as np
 from numpy.testing import assert_almost_equal
 
 import lmoments3 as lm
@@ -9,10 +10,12 @@ from lmoments3.tests import DistributionTestCase
 
 class TestLmoments(unittest.TestCase):
     def test_samlmu(self):
-        testdata = [2.0, 3.0, 4.0, 2.4, 5.5, 1.2, 5.4, 2.2, 7.1, 1.3, 1.5]
+        testdata = np.array([2.0, 3.0, 4.0, 2.4, 5.5, 1.2, 5.4, 2.2, 7.1, 1.3, 1.5])
         expected = [3.23636364, 1.14181818, 0.27388535, 0.02335456, -0.04246285]
         result = lm.lmom_ratios(testdata)
         assert_almost_equal(result, expected)
+        # Ensure input was not modified by `lmom_ratios`.
+        assert all(testdata[-3:] == [7.1, 1.3, 1.5])
 
     def test_aic(self):
         data = [